Move debugging code out of kernel
