+++ src/interactive-window/debugger/helper.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 'use strict';
-import { ISourceMapRequest } from '../../kernels/debugger/types';
+import { ISourceMapRequest } from '../../notebooks/debugger/debuggingTypes';
 import { getFilePath } from '../../platform/common/platform/fs-paths';
 import { IFileGeneratedCodes } from '../editor-integration/types';
