+++ src/interactive-window/debugger/jupyter/debugCellControllers.ts
@@ -3,10 +3,10 @@
 
 import { DebugProtocolMessage, NotebookCell } from 'vscode';
 import { DebugProtocol } from 'vscode-debugprotocol';
-import { IDebuggingDelegate, IKernelDebugAdapter } from '../../../kernels/debugger/types';
-import { DebuggingTelemetry } from '../../../kernels/debugger/constants';
+import { IDebuggingDelegate, IKernelDebugAdapter } from '../../../notebooks/debugger/debuggingTypes';
+import { DebuggingTelemetry } from '../../../notebooks/debugger/constants';
 import { IKernel } from '../../../kernels/types';
-import { cellDebugSetup } from '../../../kernels/debugger/helper';
+import { cellDebugSetup } from '../../../notebooks/debugger/helper';
 import { createDeferred } from '../../../platform/common/utils/async';
 import { sendTelemetryEvent } from '../../../telemetry';
 import { getInteractiveCellMetadata } from '../../helpers';
