+++ src/notebooks/debugger/debugLocationTracker.ts
@@ -3,7 +3,7 @@
 'use strict';
 import { DebugAdapterTracker, Event, EventEmitter } from 'vscode';
 import { DebugProtocol } from 'vscode-debugprotocol';
-import { IDebugLocation } from './types';
+import { IDebugLocation } from './debuggingTypes';
 
 // When a python debugging session is active keep track of the current debug location
 export class DebugLocationTracker implements DebugAdapterTracker {
